Step through wrappers more easily
It is quite hard to determine what is an automatically generated wrapper
and what isn't. Thus, we use the heuristic that functions defined on the
same line/file will automatically be stepped into, even if stepping in
was not explicitly requested. This feels pretty intuitive, but could
perhaps lead to false positives (E.g. if methods to some function are defined
in a loop). We'll see how this goes. If it doesn't work, we may need a more
reliable way to detected implicitly generated wrappers.

Fixes #25